CLEANUP: Suppressed warning where it is necessary. Otherwise, treat warnings as errors for the core projects.

+++ collection.c
@@ -871,7 +871,7 @@ fiftyoneDegreesCollection* fiftyoneDegreesCollectionCreateFromFile(
 	}
 
 	if (result == NULL || (
-		(int)result->count == config->loaded &&
+		result->count == config->loaded &&
 		(long)result->size < (long)(ftell(file) - header.startPosition))) {
 
 		// Create the next collection if one is needed.
